Use uint32_t for NAL offset and size
This is enough for the offsets/sizes within a single buffer, and
prevents miscommunication between 64-bit and 32-bit processes.

Bug: 27785687

+++ media/libstagefright/include/avc_utils.h
@@ -37,8 +37,8 @@ enum {
 };
 
 struct NALPosition {
-    size_t nalOffset;
-    size_t nalSize;
+    uint32_t nalOffset;
+    uint32_t nalSize;
 };
 
 // Optionally returns sample aspect ratio as well.
